third_party/WebKit/LayoutTests/fast/css/fontfaceset-loadingdone-expected.txt

Test that onloadingdone is fired after style recalculation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

loadingdone
PASS window.getComputedStyle(testDiv, 0).width is "48px"
ready promise fulfilled
PASS window.getComputedStyle(testDiv, 0).width is "48px"
PASS successfullyParsed is true

TEST COMPLETE

aaa
